requirements.txt
trimesh==3.22.2
pyrender==0.1.45
PyMCubes==0.1.4
lpips-jax==0.1.0

numpy==1.24.4
matplotlib==3.5.3

jax==0.4.9
jaxlie==1.3.1
jaxtyping==0.2.13
tyro
einops==0.6.0
opencv-python==4.7.0.72
opencv-stubs==0.0.8
scikit-learn
optax==0.1.5
imageio==2.27.0
tqdm
moviepy
rich
dm-pix==0.4.0
beautifultable==1.1.0

viser>=0.1.0

pyright==1.1.324

dacite==1.8.0
h5py==3.8.0

tabulate==0.9.0

jinja2==3.1.2

pysdf==0.1.8

git+https://github.com/brentyi/fifteen.git@a4477ef583ff4adb89a74931b487e69885adc00c
